Drop support for _MOTIF_WM_HINTS for NORMAL type windows
This atom was only used for its decorations prop. The
decorations prop is inted to tell the wm that a window
wants to draw its decorations client side and is used
in this capacity by gtk3 gnome applications as well as
things like xamp. However its usage in hildon can really
only be undesireable as it causes matchbox to size the
applicaiton to be fullscreen while still drawing the hildon
bar above the applicaiton. Except for dialogs where the
property is used to for windows that legitametly dont need
decoration like HIM.